Crate google-manager1_beta2

This documentation was generated from manager crate version 0.0.1. The original source code can be found on github.

Features

Handle the following Resources with ease from the central hub ...

deployments (delete, get, insert and list)

templates (delete, get, insert and list)

Everything else about the manager v1_beta2 API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.deployments().delete(...).doit()
let r = hub.deployments().get(...).doit()
let r = hub.deployments().insert(...).doit()
let r = hub.deployments().list(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-manager1_beta2 = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate "google-manager1_beta2" as manager1_beta2;
use manager1_beta2::cmn::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use manager1_beta2::Manager;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Manager::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.deployments().list("projectId", "region")
             .page_token("kasd")
             .max_results(-22)
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Uploads and Downlods

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Structs

AccessConfig
A Compute Engine network accessConfig. Identical to the accessConfig on corresponding Compute Engine resource.

This type is not used in any activity, and only used as part of another schema.

Action
An Action encapsulates a set of commands as a single runnable module with additional information needed during run-time.

This type is not used in any activity, and only used as part of another schema.

AllowedRule
An allowed port resource.

This type is not used in any activity, and only used as part of another schema.

AutoscalingModule
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

AutoscalingModuleStatus
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

DeployState
[Output Only] The current state of a replica or module.

This type is not used in any activity, and only used as part of another schema.

Deployment
A deployment represents a physical instantiation of a Template.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete deployments (none)

get deployments (response)

insert deployments (request|response)

list deployments (none)

DeploymentDeleteMethodBuilder

DeploymentGetMethodBuilder

DeploymentInsertMethodBuilder

DeploymentListMethodBuilder

DeploymentMethodsBuilder
A builder providing access to all methods supported on deployment resources. It is not used directly, but through the Manager hub.

DeploymentsListResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list deployments (response)

DiskAttachment
How to attach a disk to a Replica.

This type is not used in any activity, and only used as part of another schema.

EnvVariable
An environment variable.

This type is not used in any activity, and only used as part of another schema.

ExistingDisk
A pre-existing persistent disk that will be attached to every Replica in the Pool.

This type is not used in any activity, and only used as part of another schema.

FirewallModule
A Firewall resource

This type is not used in any activity, and only used as part of another schema.

FirewallModuleStatus
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

HealthCheckModule
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

HealthCheckModuleStatus
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

LbModule
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

LbModuleStatus
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Manager
Central instance to access all Manager related resource activities

Metadata
A Compute Engine metadata entry. Identical to the metadata on the corresponding Compute Engine resource.

This type is not used in any activity, and only used as part of another schema.

MetadataItem
A Compute Engine metadata item, defined as a key:value pair. Identical to the metadata on the corresponding Compute Engine resource.

This type is not used in any activity, and only used as part of another schema.

Module
A module in a configuration. A module represents a single homogeneous, possibly replicated task.

This type is not used in any activity, and only used as part of another schema.

ModuleStatus
[Output Only] Aggregate status for a module.

This type is not used in any activity, and only used as part of another schema.

NetworkInterface
A Compute Engine NetworkInterface resource. Identical to the NetworkInterface on the corresponding Compute Engine resource.

This type is not used in any activity, and only used as part of another schema.

NetworkModule
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

NetworkModuleStatus
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

NewDisk
A Persistent Disk resource that will be created and attached to each Replica in the Pool. Each Replica will have a unique persistent disk that is created and attached to that Replica.

This type is not used in any activity, and only used as part of another schema.

NewDiskInitializeParams
Initialization parameters for creating a new disk.

This type is not used in any activity, and only used as part of another schema.

ParamOverride
A specification for overriding parameters in a Template that corresponds to the Deployment.

This type is not used in any activity, and only used as part of another schema.

ReplicaPoolModule
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

ReplicaPoolModuleStatus
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

ReplicaPoolParams
Configuration information for a ReplicaPools resource. Specifying an item within will determine the ReplicaPools API version used for a ReplicaPoolModule. Only one may be specified.

This type is not used in any activity, and only used as part of another schema.

ReplicaPoolParamsV1Beta1
Configuration information for a ReplicaPools v1beta1 API resource. Directly maps to ReplicaPool InitTemplate.

This type is not used in any activity, and only used as part of another schema.

ServiceAccount
A Compute Engine service account, identical to the Compute Engine resource.

This type is not used in any activity, and only used as part of another schema.

Tag
A Compute Engine Instance tag, identical to the tags on the corresponding Compute Engine Instance resource.

This type is not used in any activity, and only used as part of another schema.

Template
A Template represents a complete configuration for a Deployment.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete templates (none)

get templates (response)

list templates (none)

insert templates (request|response)

TemplateDeleteMethodBuilder

TemplateGetMethodBuilder

TemplateInsertMethodBuilder

TemplateListMethodBuilder

TemplateMethodsBuilder
A builder providing access to all methods supported on template resources. It is not used directly, but through the Manager hub.

TemplatesListResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list templates (response)

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.