Module google_manager1_beta2::api

Structs

AccessConfig
A Compute Engine network accessConfig. Identical to the accessConfig on corresponding Compute Engine resource.

Action
An Action encapsulates a set of commands as a single runnable module with additional information needed during run-time.

AllowedRule
An allowed port resource.

AutoscalingModule
There is no detailed description.

AutoscalingModuleStatus
There is no detailed description.

DeployState
[Output Only] The current state of a replica or module.

Deployment
A deployment represents a physical instantiation of a Template.

DeploymentDeleteCall
A builder for the delete method supported by a deployment resource. It is not used directly, but through a DeploymentMethods instance.

DeploymentGetCall
A builder for the get method supported by a deployment resource. It is not used directly, but through a DeploymentMethods instance.

DeploymentInsertCall
A builder for the insert method supported by a deployment resource. It is not used directly, but through a DeploymentMethods instance.

DeploymentListCall
A builder for the list method supported by a deployment resource. It is not used directly, but through a DeploymentMethods instance.

DeploymentMethods
A builder providing access to all methods supported on deployment resources. It is not used directly, but through the Manager hub.

DeploymentsListResponse
There is no detailed description.

DiskAttachment
How to attach a disk to a Replica.

EnvVariable
An environment variable.

ExistingDisk
A pre-existing persistent disk that will be attached to every Replica in the Pool.

FirewallModule
A Firewall resource

FirewallModuleStatus
There is no detailed description.

HealthCheckModule
There is no detailed description.

HealthCheckModuleStatus
There is no detailed description.

LbModule
There is no detailed description.

LbModuleStatus
There is no detailed description.

Manager
Central instance to access all Manager related resource activities

Metadata
A Compute Engine metadata entry. Identical to the metadata on the corresponding Compute Engine resource.

MetadataItem
A Compute Engine metadata item, defined as a key:value pair. Identical to the metadata on the corresponding Compute Engine resource.

Module
A module in a configuration. A module represents a single homogeneous, possibly replicated task.

ModuleStatus
[Output Only] Aggregate status for a module.

NetworkInterface
A Compute Engine NetworkInterface resource. Identical to the NetworkInterface on the corresponding Compute Engine resource.

NetworkModule
There is no detailed description.

NetworkModuleStatus
There is no detailed description.

NewDisk
A Persistent Disk resource that will be created and attached to each Replica in the Pool. Each Replica will have a unique persistent disk that is created and attached to that Replica.

NewDiskInitializeParams
Initialization parameters for creating a new disk.

ParamOverride
A specification for overriding parameters in a Template that corresponds to the Deployment.

ReplicaPoolModule
There is no detailed description.

ReplicaPoolModuleStatus
There is no detailed description.

ReplicaPoolParams
Configuration information for a ReplicaPools resource. Specifying an item within will determine the ReplicaPools API version used for a ReplicaPoolModule. Only one may be specified.

ReplicaPoolParamsV1Beta1
Configuration information for a ReplicaPools v1beta1 API resource. Directly maps to ReplicaPool InitTemplate.

ServiceAccount
A Compute Engine service account, identical to the Compute Engine resource.

Tag
A Compute Engine Instance tag, identical to the tags on the corresponding Compute Engine Instance resource.

Template
A Template represents a complete configuration for a Deployment.

TemplateDeleteCall
A builder for the delete method supported by a template resource. It is not used directly, but through a TemplateMethods instance.

TemplateGetCall
A builder for the get method supported by a template resource. It is not used directly, but through a TemplateMethods instance.

TemplateInsertCall
A builder for the insert method supported by a template resource. It is not used directly, but through a TemplateMethods instance.

TemplateListCall
A builder for the list method supported by a template resource. It is not used directly, but through a TemplateMethods instance.

TemplateMethods
A builder providing access to all methods supported on template resources. It is not used directly, but through the Manager hub.

TemplatesListResponse
There is no detailed description.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.