Trait google-manager1_beta2::cmn::Delegate

pub trait Delegate {
    fn begin(&mut self, MethodInfo) { ... }
    fn http_error(&mut self, &HttpError) -> Retry { ... }
    fn api_key(&mut self) -> Option<String> { ... }
    fn token(&mut self) -> Option<Token> { ... }
    fn http_failure(&mut self, _: &Response, JsonServerError) -> Retry { ... }
    fn pre_request(&mut self) { ... }
    fn finished(&mut self) { ... }
}

A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

It contains methods to deal with all common issues, as well with the ones related to uploading media

Provided Methods

fn begin(&mut self, MethodInfo)

Called at the beginning of any API request. The delegate should store the method information if he is interesting in knowing more context when further calls to it are made. The matching finished() call will always be made, no matter whether or not the API request was sucessfull. That way, the delgate may easily maintain a clean state between various API calls.

fn http_error(&mut self, &HttpError) -> Retry

Called whenever there is an HttpError, usually if there are network problems.

Return retry information.

fn api_key(&mut self) -> Option<String>

Called whenever there is the need for your applications API key after the official authenticator implementation didn't provide one, for some reason. If this method returns None as well, the underlying operation will fail

fn token(&mut self) -> Option<Token>

Called whenever the Authenticator didn't yield a token. The delegate may attempt to provide one, or just take is a general information about the pending impending failure

fn http_failure(&mut self, _: &Response, JsonServerError) -> Retry

Called whenever the http request returns with a non-success status code. This can involve authentication issues, or anything else that very much depends on the used API method. The delegate should check the status, header and decoded json error to decide whether to retry or not. In the latter case, the underlying call will fail.

fn pre_request(&mut self)

Called prior to sending the main request of the given method. It can be used to time the call or to print progress information. It's also useful as you can be sure that a request will definitely be made.

fn finished(&mut self)

Called before the API request method returns, in every case. It can be used to clean up internal state between calls to the API. This call always has a matching call to begin(...).