Struct google_manager1_beta2::Module

pub struct Module {
    pub lb_module: Option<LbModule>,
    pub replica_pool_module: Option<ReplicaPoolModule>,
    pub firewall_module: Option<FirewallModule>,
    pub health_check_module: Option<HealthCheckModule>,
    pub autoscaling_module: Option<AutoscalingModule>,
    pub network_module: Option<NetworkModule>,
    pub type_: Option<String>,
}

A module in a configuration. A module represents a single homogeneous, possibly replicated task.

This type is not used in any activity, and only used as part of another schema.

Fields
lb_module: Option<LbModule>

no description provided

replica_pool_module: Option<ReplicaPoolModule>

no description provided

firewall_module: Option<FirewallModule>

no description provided

health_check_module: Option<HealthCheckModule>

no description provided

autoscaling_module: Option<AutoscalingModule>

no description provided

network_module: Option<NetworkModule>

no description provided

type_: Option<String>

The type of this module. Valid values ("AUTOSCALING", "FIREWALL", "HEALTH_CHECK", "LOAD_BALANCING", "NETWORK", "REPLICA_POOL")