Struct google_manager1_beta2::ReplicaPoolModule

pub struct ReplicaPoolModule {
    pub health_checks: Option<Vec<String>>,
    pub replica_pool_params: Option<ReplicaPoolParams>,
    pub resource_view: Option<String>,
    pub num_replicas: Option<i32>,
    pub env_variables: Option<HashMap<String, EnvVariable>>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields

health_checks

The Health Checks to configure for the ReplicaPoolModule

replica_pool_params

Information for a ReplicaPoolModule.

resource_view

[Output Only] The name of the Resource View associated with a ReplicaPoolModule. This field will be generated by the service.

num_replicas

Number of replicas in this module.

env_variables

A list of environment variables.