Struct google_manager1_beta2::ReplicaPoolParamsV1Beta1

pub struct ReplicaPoolParamsV1Beta1 {
    pub disks_to_create: Option<Vec<NewDisk>>,
    pub auto_restart: Option<bool>,
    pub can_ip_forward: Option<bool>,
    pub description: Option<String>,
    pub zone: Option<String>,
    pub tags: Option<Tag>,
    pub on_host_maintenance: Option<String>,
    pub disks_to_attach: Option<Vec<ExistingDisk>>,
    pub machine_type: Option<String>,
    pub base_instance_name: Option<String>,
    pub init_action: Option<String>,
    pub metadata: Option<Metadata>,
    pub service_accounts: Option<Vec<ServiceAccount>>,
    pub network_interfaces: Option<Vec<NetworkInterface>>,
}

Configuration information for a ReplicaPools v1beta1 API resource. Directly maps to ReplicaPool InitTemplate.

This type is not used in any activity, and only used as part of another schema.

Fields

disks_to_create

A list of Disk resources to create and attach to each Replica in the Pool. Currently, you can only define one disk and it must be a root persistent disk. Note that Replica Pool will create a root persistent disk for each replica.

auto_restart

Whether these replicas should be restarted if they experience a failure. The default value is true.

can_ip_forward

Enables IP Forwarding

description

An optional textual description of the resource.

zone

The zone for this ReplicaPool.

tags

A list of tags to apply to the Google Compute Engine instance to identify resources.

on_host_maintenance

no description provided

disks_to_attach

A list of existing Persistent Disk resources to attach to each replica in the pool. Each disk will be attached in read-only mode to every replica.

machine_type

The machine type for this instance. Either a complete URL, or the resource name (e.g. n1-standard-1).

base_instance_name

The base name for instances within this ReplicaPool.

init_action

Name of the Action to be run during initialization of a ReplicaPoolModule.

metadata

The metadata key/value pairs assigned to this instance.

service_accounts

A list of Service Accounts to enable for this instance.

network_interfaces

A list of network interfaces for the instance. Currently only one interface is supported by Google Compute Engine.