Struct google_manager1_beta2::Template

pub struct Template {
    pub modules: Option<HashMap<String, Module>>,
    pub description: Option<String>,
    pub actions: Option<HashMap<String, Action>>,
    pub name: Option<String>,
}

A Template represents a complete configuration for a Deployment.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete templates (none)

get templates (response)

list templates (none)

insert templates (request|response)

Fields

modules

A list of modules for this Template.

description

A user-supplied description of this Template.

actions

Action definitions for use in Module intents in this Template.

name

Name of this Template. The name must conform to the expression: [a-zA-Z0-9-_]{1,64}