Struct google_manager1_beta2::Deployment

pub struct Deployment {
    pub state: Option<DeployState>,
    pub description: Option<String>,
    pub template_name: Option<String>,
    pub overrides: Option<Vec<ParamOverride>>,
    pub creation_date: Option<String>,
    pub modules: Option<HashMap<String, ModuleStatus>>,
    pub name: Option<String>,
}

A deployment represents a physical instantiation of a Template.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete deployments (none)

get deployments (response)

insert deployments (request|response)

list deployments (none)

Fields

state

[Output Only] Current status of this deployment.

description

A user-supplied description of this Deployment.

template_name

The name of the Template on which this deployment is based.

overrides

The set of parameter overrides to apply to the corresponding Template before deploying.

creation_date

[Output Only] The time when this deployment was created.

modules

[Output Only] List of status for the modules in this deployment.

name

Name of this deployment. The name must conform to the following regular expression: [a-zA-Z0-9-_]{1,64}