Struct google_manager1_beta2::Manager

pub struct Manager<C, A> {
    // some fields omitted
}

Central instance to access all Manager related resource activities

Examples

Instantiate a new hub

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_manager1_beta2 as manager1_beta2;
use manager1_beta2::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use manager1_beta2::Manager;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Manager::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.deployments().list("projectId", "region")
             .page_token("kasd")
             .max_results(-22)
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Methods

impl<'a, C, A> Manager<C, A> where C: BorrowMut<Client>, A: GetToken

fn new(client: C, authenticator: A) -> Manager<C, A>

fn deployments(&'a self) -> DeploymentMethods<'a, C, A>

fn templates(&'a self) -> TemplateMethods<'a, C, A>

fn user_agent(&mut self, agent_name: String) -> String

Set the user-agent header field to use in all requests to the server. It defaults to google-api-rust-client/0.1.11.

Returns the previously set user-agent.